refactor: Split realm query methods in DriveInfosManager+Query
